requirements.txt
Flask==2.0.3
Flask-SocketIO==5.1.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==0.7.1
Flask-WTF==0.14.3
pyDH==1.4.2
rsa==4.7.2
pycryptodome==3.10.4
WTForms==3.0.1
python-dotenv==0.19.2